src/requirements.txt
requests>=2.31.0
langfuse>=2.0.0
openai>=1.40.0
pydantic>=1.10.13,<3
supabase>=2.0.0
python-docx>=1.1.0
google-genai
groq>=0.9.0
parallel-web>=0.4.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
opentelemetry-instrumentation-google-genai>=0.6b0
